file エンティティ分割
