+++ src/entities/file/mod.rs
@@ -1,0 +1,11 @@
+mod file;
+mod file_id;
+mod file_key;
+mod file_size;
+mod mime_type;
+
+pub use file::File;
+pub use file_id::FileId;
+pub use file_key::FileKey;
+pub use file_size::{FileSize, FileSizeTryFromError};
+pub use mime_type::{MimeType, MimeTypeTryFromError};
